Complete rework
This patch changes the functions 'queue_init()' and 'queue_dest()' to
'queue_create()' and 'queue_destroy()', respectively. See the doc
comments. In addition, the structure was renamed to 'qarray', and all
API functions were renamed from 'queue_<...>' to 'qarray_<...>' for
consistency. The test code was updated to reflect these changes.